64: Fix return address relocation in CheckStackGuardState
Port of PPC64 fix: https://github.com/andrewlow/v8ppc/commit/bf98316b114a1455e3dfbccc2ddd5413e9631211

+++ src/s390/regexp-macro-assembler-s390.cc
@@ -1282,7 +1282,7 @@ int RegExpMacroAssemblerS390::CheckStackGuardState(Address* return_address,
   MaybeObject* result = Execution::HandleStackGuardInterrupt(isolate);
 
   if (*code_handle != re_code) {  // Return address no longer valid
-    int delta = code_handle->address() - re_code->address();
+    intptr_t delta = code_handle->address() - re_code->address();
     // Overwrite the return address on the stack.
     *return_address += delta;
   }
